agenteV1/landing_page_agent/requirements.txt
# Dependencias del proyecto
Flask==2.3.2
openai==0.27.0
requests==2.31.0
google-generativeai==0.3.0
colorama==0.4.6